added warrior abilities and stun status, fixed multi target abilties to work fewer enemies than their max

+++ Assets/Scripts/PlayerScripts/PlayerAbilities/Defensive/MageDefensive/Healing/DoubleHealS.cs
@@ -8,7 +8,7 @@ public class DoubleHealS : DefensiveAbility {
 	{
 		bool success = false;
 		foreach (Player p in targets) {
-			int heal = (p.GetMaxHealth () * (1 / 3));
+			int heal = (int)(p.GetMaxHealth () * (1 / 3));
 			success = p.RestoreHealth(heal);
 		}
 		return success;
